refactor(auth): centralize auth configuration in lib/auth
Move authentication logic to a centralized authOptions object in lib/auth.js to improve maintainability and reduce code duplication. Update API routes to use the new authOptions object for session management.
