+++ src/app/api/auth/[...nextauth]/route.js
@@ -1,72 +1,6 @@
 import NextAuth from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-import { MongoClient } from 'mongodb';
-import bcrypt from 'bcryptjs';
+import { authOptions } from '@/lib/auth';
 
-const handler = NextAuth({
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        username: { label: "Username", type: "text" },
-        password: { label: "Password", type: "password" }
-      },
-      async authorize(credentials) {
-        try {
-          const client = await MongoClient.connect(process.env.MONGODB_URI);
-          const db = client.db();
-          
-          const user = await db.collection('users').findOne({
-            username: credentials.username
-          });
-
-          if (!user) {
-            client.close();
-            return null;
-          }
-
-          const isValid = await bcrypt.compare(credentials.password, user.password);
-
-          if (!isValid) {
-            client.close();
-            return null;
-          }
-
-          client.close();
-          return {
-            id: user._id.toString(),
-            name: user.username,
-            role: user.role
-          };
-        } catch (error) {
-          console.error('Error in authorize:', error);
-          return null;
-        }
-      }
-    })
-  ],
-  session: {
-    strategy: 'jwt',
-  },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.role = user.role;
-        token.id = user.id;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (token) {
-        session.user.role = token.role;
-        session.user.id = token.id;
-      }
-      return session;
-    }
-  },
-  pages: {
-    signIn: '/login',
-  },
-});
+const handler = NextAuth(authOptions);
 
 export { handler as GET, handler as POST };
